Clean up user controller by adding lib folder and extracting complexity
